requirements.txt
numpy==1.16.2
pdoc3==0.5.3
Pillow==5.4.1
PyOpenGL==3.1.0
pyrr==0.10.2
PySDL2==0.9.6
read==2018.11.20
